requirements.txt
agate==1.6.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bitarray==2.7.3
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.14.6
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
cookiecutter==2.1.1
cryptography==40.0.2
cssselect==1.2.0
cytoolz==0.12.1
dataclasses-json==0.5.7
dbt-core==0.19.2
dbt-mysql==0.19.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
docker==6.0.1
docopt==0.6.2
duckduckgo-search==2.8.5
ed25519==1.5
entrypoints==0.4
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.3.2
Flask-Cors==3.0.10
Flask-WTF==1.1.1
fqdn==1.5.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
hexbytes==0.3.0
hologram==0.0.13
httplib2==0.22.0
idna==3.4
importlib-metadata==6.3.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
json-rpc==1.14.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
langchain==0.0.138
leather==0.3.4
Logbook==1.5.3
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mashumaro==2.0
matplotlib-inline==0.1.6
minimal-snowplow-tracker==0.0.2
mistune==2.0.5
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parsedatetime==2.6
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.1
pipreqs==0.4.11
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.0
pytest-cookies==0.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
PyYAML==6.0
pyzmq==24.0.1
readability-lxml==0.8.1
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==3.0.0
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
sqlparse==0.3.1
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
tiktoken==0.3.3
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
waitress==2.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
WTForms==3.0.1
yarg==0.1.9
yarl==1.8.2
zini==1.1.0
zipp==3.15.0
